requirements.txt
Flask==3.0.3
reportlab==4.2.2
pypdf==5.1.0
Pillow==10.4.0
pdfminer.six==20231228
requests==2.32.3
python-dateutil==2.9.0.post0
python-stdnum==1.19
